requirements.txt
gdown==3.12.2
opencv-python==4.4.0.46
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
wandb~=0.10.10
tqdm~=4.51.0
numpy~=1.19.4
Pillow~=8.0.1
